move env-graph within src to make turbo caching happy
