+++ packages/varlock/src/env-graph/lib/config-item.ts
