+++ packages/varlock/src/env-graph/lib/graph-utils.ts
