+++ packages/varlock/src/env-graph/lib/simple-queue.ts
