+++ packages/varlock/src/env-graph/lib/type-generation.ts
