+++ packages/varlock/src/env-graph/test/plugins/test-plugin-conflict-1/package.json
