+++ packages/varlock/src/env-graph/test/plugins/test-plugin-conflict-1/plugin.js
@@ -1,4 +1,4 @@
-/// <reference path="../../../../src/plugin-lib.ts" />
+/// <reference path="../../../../plugin-lib.ts" />
 
 plugin.registerResolverFunction({
   name: 'conflict',
