+++ packages/varlock/src/env-graph/test/plugins/test-plugin-conflict-2/package.json
