+++ packages/varlock/src/env-graph/test/plugins/test-plugin-version-conflict/package.json
