+++ packages/varlock/src/env-graph/test/plugins/test-plugin-version-conflict/plugin.js
@@ -1,4 +1,4 @@
-/// <reference path="../../../../src/plugin-lib.ts" />
+/// <reference path="../../../../plugin-lib.ts" />
 
 plugin.registerResolverFunction({
   name: 'noop',
