+++ packages/varlock/src/lib/load-graph.ts
@@ -1,4 +1,4 @@
-import { loadEnvGraph } from '../../env-graph';
+import { loadEnvGraph } from '../env-graph';
 
 export async function loadVarlockEnvGraph(opts?: {
   currentEnvFallback?: string,
