+++ packages/varlock/src/plugin-lib.ts
@@ -1,5 +1,5 @@
-import type { VarlockPlugin } from '../env-graph/lib/plugins';
-export type { Resolver } from '../env-graph/lib/resolver';
+import type { VarlockPlugin } from './env-graph/lib/plugins';
+export type { Resolver } from './env-graph/lib/resolver';
 
 // we must inject our varlock code via the global scope
 // because of how we dynamically import the plugin code
