+++ packages/varlock/tsconfig.json
@@ -18,7 +18,6 @@
   },
   "include": [
     "src/**/*.ts",
-    "env-graph/**/*.ts",
     "*.ts",
   ],
   "exclude": ["src/plugin-lib.ts"]
